# back/src/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "MeLi Challenge",
    "contact": {
      "email": "danielalejandro.mn19@gmail.com",
      "name": "Daniel Alejandro Melo Nuvan"
    }
  },
  "host": "example.com",
  "basePath": "localhost:5000",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1.1/register": {
      "post": {
        "summary": "Register",
        "tags": [
          "Auth"
        ],
        "operationId": "Register",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/v1.1/login": {
      "get": {
        "summary": "Login",
        "tags": [
          "Auth"
        ],
        "operationId": "Login",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/loginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": {
                "jwt": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJleHAiOjE2OTg5NzE5MzJ9.bUXASIXciKC584Y9otLcfbGn0t01N2SXFmbqnFyFLo4",
                "refresh": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJleHAiOjE2OTkwNTY1MzJ9.QxEqVUdImZUviZrhMwRteN3LqTO-qH8Es0H4GpfEZCY"
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/v1.1/refresh": {
      "get": {
        "summary": "refresh",
        "tags": [
          "Auth"
        ],
        "operationId": "refresh",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": true,
            "type": "string",
            "description": "The refresh token"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": {
                "jwt": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJleHAiOjE2OTg5NzIxNTN9.1UhbiEby1HOTWvb6_Fcz740sPOBcwtuS4bp1vOYPziY",
                "refresh": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJleHAiOjE2OTkwNTY3NTN9.QntzeFZy7GtognJquCyXfq1DaCLDr2vAV6twMUwSgAk"
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/v1.1/providers": {
      "get": {
        "summary": "Get all",
        "tags": [
          "Providers"
        ],
        "operationId": "Getall",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": [
                {
                  "country": "col",
                  "created_at": "Thu, 02 Nov 2023 20:02:38 GMT",
                  "description": "test description",
                  "id": 1,
                  "name": "test provider 2",
                  "updated_at": "Thu, 02 Nov 2023 20:59:33 GMT"
                }
              ]
            },
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "summary": "Create",
        "tags": [
          "Providers"
        ],
        "operationId": "Create",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/v1.1/providers/{provider_id}": {
      "get": {
        "summary": "Get one id",
        "tags": [
          "Providers"
        ],
        "operationId": "Getoneid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": {
                "country": "Thu, 02 Nov 2023 20:02:38 GMT",
                "created_at": "Thu, 02 Nov 2023 20:59:33 GMT",
                "description": "test description",
                "id": 1,
                "name": "test provider 2"
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete",
        "tags": [
          "Providers"
        ],
        "operationId": "Delete",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "summary": "Update",
        "tags": [
          "Providers"
        ],
        "operationId": "Update",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/v1.1/roles": {
      "get": {
        "summary": "Get all",
        "tags": [
          "Roles"
        ],
        "operationId": "Getall1",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": [
                {
                  "created_at": "Thu, 02 Nov 2023 19:38:16 GMT",
                  "id": 1,
                  "name": "ADMIN",
                  "updated_at": "Thu, 02 Nov 2023 19:38:16 GMT"
                },
                {
                  "created_at": "Thu, 02 Nov 2023 21:46:12 GMT",
                  "id": 2,
                  "name": "TEST_ROLE",
                  "updated_at": "Thu, 02 Nov 2023 21:46:12 GMT"
                }
              ]
            },
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "summary": "Create",
        "tags": [
          "Roles"
        ],
        "operationId": "PostCreate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateRequest1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/v1.1/roles/{role_id}": {
      "put": {
        "summary": "Update",
        "tags": [
          "Roles"
        ],
        "operationId": "PutUpdate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest1"
            }
          },
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete",
        "tags": [
          "Roles"
        ],
        "operationId": "Delete1",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/v1.1/users/{user_id}/roles": {
      "get": {
        "summary": "get current user roles",
        "tags": [
          "users"
        ],
        "operationId": "getcurrentuserroles",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": {
                "created_at": "Thu, 02 Nov 2023 19:38:36 GMT",
                "email": "test@test.com",
                "id": 1,
                "name": "test",
                "roles": [
                  "ADMIN"
                ],
                "updated_at": "Thu, 02 Nov 2023 19:38:36 GMT"
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "summary": "add role to user",
        "tags": [
          "users"
        ],
        "operationId": "addroletouser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/addroletouserrequest"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/v1.1/profile": {
      "get": {
        "summary": "get current user profile",
        "tags": [
          "users"
        ],
        "operationId": "getcurrentuserprofile",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "default": ""
            },
            "examples": {
              "text/plain": {
                "created_at": "Thu, 02 Nov 2023 19:38:36 GMT",
                "email": "test@test.com",
                "id": 1,
                "name": "test",
                "updated_at": "Thu, 02 Nov 2023 19:38:36 GMT"
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "RegisterRequest": {
      "title": "RegisterRequest",
      "example": {
        "email": "test@test.com",
        "name": "test",
        "password": "testPassword"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "name",
        "password"
      ]
    },
    "loginRequest": {
      "title": "loginRequest",
      "example": {
        "email": "test@test.com",
        "password": "testPassword"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "CreateRequest": {
      "title": "CreateRequest",
      "example": {
        "name": "test provider",
        "description": "test description",
        "country": "col"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description",
        "country"
      ]
    },
...
  },
  "security": [],
...
}